chore: move user related tests to UserTests
